Checks for the unique Name on the ContactAdd Moved userData from ContactsList to the App.
Added Route check for the userData, only the routes ADD EDIT VIEW to be accessed if there is userData.

Created check on the ContactAdd to only continue if Name is unique - it will be used like its unique id later on.

Added Error Popup to the ContactAdd component.

Edited some h1's to be centered.